Mapping for int_div, int_rem and bunch of other int notions (#6)
This PR proposes multiple alignments, mostly for integers, and sometimes for some real notions that are required as dependencies.

I also proved some util lemmas for `COND` and I don't know where to put them. Maybe the code should be refactored to use other `COND` lemmas instead.

----------------------

This PR also introduces some handy tactics such as `ext` to perform `apply fun_ext` followed by `intro` and a tactic `align_ε` to prove equalities where the RHS is `ε` applied to an arbitrary number of arguments, and getting rid of those that are useless like tags (then the proof doesn't have to even mention it).

----------------------

The work on `Z.lcm` was a collaboration with @NotBad4U.